rename SolrFirstPassGroupingCollector to SkipSecondPassFirstPassGroupingCollector to reflect usage; make SolrSearchGroup an inner class of it;
