--- solr/core/src/java/org/apache/solr/search/grouping/SolrSearchGroup.java
@@ -1,65 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.solr.search.grouping;
-
-import org.apache.lucene.search.grouping.SearchGroup;
-
-public class SolrSearchGroup<T> extends SearchGroup<T> {
-
-  public int topDocLuceneId;
-  public float topDocScore;
-  public Object topDocSolrId;
-
-  @Override
-  protected MergedGroup<T> newMergedGroup() {
-    SolrMergedGroup<T> mergedGroup = new SolrMergedGroup<>(this.groupValue);
-    mergedGroup.topDocScore = this.topDocScore;
-    mergedGroup.topDocSolrId = this.topDocSolrId;
-    return mergedGroup;
-  }
-
-  private static class SolrMergedGroup<T> extends SearchGroup.MergedGroup<T> {
-
-    public float topDocScore;
-    public Object topDocSolrId;
-
-    public SolrMergedGroup(T groupValue) {
-      super(groupValue);
-    }
-
-    @Override
-    protected SearchGroup<T> newSearchGroup() {
-      return new SolrSearchGroup<T>();
-    }
-
-    @Override
-    protected void fillSearchGroup(SearchGroup<T> searchGroup) {
-      super.fillSearchGroup(searchGroup);
-      ((SolrSearchGroup<T>)searchGroup).topDocScore = this.topDocScore;
-      ((SolrSearchGroup<T>)searchGroup).topDocSolrId = this.topDocSolrId;
-    }
-
-    @Override
-    public void update(SearchGroup<T> searchGroup) {
-      super.update(searchGroup);
-      this.topDocScore = ((SolrSearchGroup<T>)searchGroup).topDocScore;
-      this.topDocSolrId = ((SolrSearchGroup<T>)searchGroup).topDocSolrId;
-    }
-
-  }
-
-}
